cloud/rec-sys-training/requirements.txt
pandas==3.0.0
numpy==2.4.2
seaborn==0.13.2
matplotlib==3.10.8
requests==2.32.5
BeautifulSoup4==4.14.3
selenium==4.40.0
loguru==0.7.3
pyarrow==23.0.1
fastparquet==2025.12.0
scikit-learn==1.8.0
nltk==3.9.3
matplotlib==3.10.8
fastapi==0.135.1
uvicorn[standard]==0.41.0
jinja2==3.1.6
python-multipart==0.0.22
pandas-gbq==0.34.0
google-cloud-bigquery==3.40.1